Move upgradeFiles function to a new file
